requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Werkzeug==2.0.3
anthropic==0.18.1
python-dotenv==1.0.1
pydantic==2.10.4
Jinja2==3.0.3
MarkupSafe==2.0.1
itsdangerous==2.0.1
click==8.1.7

# Resume parser dependencies
pdfminer.six==20250327
python-docx==1.1.2
spacy==3.8.5
nltk==3.9.1